+++ internal/pkg/client/oci/pull.go
@@ -21,20 +21,21 @@ import (
 	"github.com/apptainer/apptainer/internal/pkg/build/oci"
 	"github.com/apptainer/apptainer/internal/pkg/cache"
 	"github.com/apptainer/apptainer/internal/pkg/util/fs"
+	"github.com/apptainer/apptainer/internal/pkg/util/ociauth"
 	buildtypes "github.com/apptainer/apptainer/pkg/build/types"
-	"github.com/apptainer/apptainer/pkg/syfs"
 	"github.com/apptainer/apptainer/pkg/sylog"
 	useragent "github.com/apptainer/apptainer/pkg/util/user-agent"
 	ocitypes "github.com/containers/image/v5/types"
 )
 
 type PullOptions struct {
-	TmpDir     string
-	OciAuth    *ocitypes.DockerAuthConfig
-	DockerHost string
-	NoHTTPS    bool
-	NoCleanUp  bool
-	Pullarch   string
+	TmpDir      string
+	OciAuth     *ocitypes.DockerAuthConfig
+	DockerHost  string
+	NoHTTPS     bool
+	NoCleanUp   bool
+	Pullarch    string
+	ReqAuthFile string
 }
 
 // pull will build a SIF image into the cache if directTo="", or a specific file if directTo is set.
@@ -47,7 +48,7 @@ func pull(ctx context.Context, imgCache *cache.Handle, directTo, pullFrom string
 	sysCtx := &ocitypes.SystemContext{
 		OCIInsecureSkipTLSVerify: opts.NoHTTPS,
 		DockerAuthConfig:         opts.OciAuth,
-		AuthFilePath:             syfs.SearchDockerConf(),
+		AuthFilePath:             ociauth.ChooseAuthFile(opts.ReqAuthFile),
 		DockerRegistryUserAgent:  useragent.Value(),
 		BigFilesTemporaryDir:     opts.TmpDir,
 	}
@@ -125,6 +126,7 @@ func convertOciToSIF(ctx context.Context, imgCache *cache.Handle, image, cachedI
 				DockerDaemonHost: opts.DockerHost,
 				ImgCache:         imgCache,
 				Arch:             opts.Pullarch,
+				ReqAuthFile:      opts.ReqAuthFile,
 			},
 		},
 	)